backend/requirements.txt
fastapi==0.101.1
pydantic==2.3.0
uvicorn==0.23.2
tortoise-orm==0.20.0
redis==5.0.0
websockets==11.0.3